pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "bonk_bot"
version = "1.1.4"
description = "Python API wrapper for bonk.io web game."

readme = "README.md"
license = {file = "LICENSE"}

authors = [{name = "Safizapi"}]
maintainers = [{name = "Safizapi"}]

dependencies = [
    "python-engineio==3.14.2",
    "python-socketio==4.6.0",
    "aiohttp==3.9.5",
    "requests==2.32.3",
    "nest-asyncio==1.6.0",
    "pymitter==0.5.1"
]
requires-python = ">= 3.8"

keywords = ["bonk", "bonk.io", "bots", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",

    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12"
]

[project.urls]
Repository = "https://github.com/Safizapi/bonk_bot"
